arb-data-prefect/requirements.txt
prefect>=3.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
requests>=2.31.0
python-dotenv>=1.0.0

